Project structure is changed: Seperate View and Controller classes. FXML is used to define the UI. Controller classes enable interaction with the GUI.
