+++ src/main/java/controller/Controller.java
@@ -1,0 +1,9 @@
+package controller;
+
+import javafx.fxml.Initializable;
+
+public interface Controller extends Initializable{
+	 
+	public void close();
+	
+}
