+++ src/main/java/mainMethod/Main.java
@@ -1,17 +1,19 @@
 package mainMethod;
 
-import userinterface.TrackingDataView;
 
 public class Main {
 
     public static void main(String[] args) {
 
-        new Thread() {
-            @Override
-            public void run() {
-                javafx.application.Application.launch(TrackingDataView.class);
-            }
-        }.start();
+//        new Thread() {
+//            @Override
+//            public void run() {
+            	javafx.application.Application.launch(ApplicationWindow.class);
+//            }
+//            
+//        }.start();
+    	
+    	
 
     }
